[ALPS09648842] vpu: Fix potential overflow issue in timestamp
[Description]
- Modified VPU_TS_NS macro in vpu_hw.c to cast timespec64_to_ns
results to uint64_t before subtraction to prevent overflow.
- Updated period calculation in vpu_main.c to cast timespec64_to_ns
results to uint64_t before subtraction to ensure accurate period
calculation.

This change addresses the overflowed constant issue reported by
Coverity.

[Test]
Build pass

MTK-Commit-Id: 86bbf9ae5b390c755dbcb3c6b8505e6e7da84f1a

CR-Id: ALPS09648842
Feature: apu_vpu_kernel

+++ drivers/misc/mediatek/apusys/vpu/4.5/vpu_hw.c
@@ -38,7 +38,7 @@
 	ktime_get_ts64(&a##_e)
 
 #define VPU_TS_NS(a) \
-	((uint64_t)(timespec64_to_ns(&a##_e) - timespec64_to_ns(&a##_s)))
+	((uint64_t)((uint64_t)timespec64_to_ns(&a##_e) - (uint64_t)timespec64_to_ns(&a##_s)))
 
 #define VPU_TS_US(a) \
 	(VPU_TS_NS(a) / 1000)
